move json schema generation to scripts/
